+++ src/main/java/leets/leenk/domain/feed/domain/entity/Feed.java
@@ -1,35 +1,36 @@
-package leets.leenk.domain.feed.domain.enitty;
+package leets.leenk.domain.feed.domain.entity;
 
 import jakarta.persistence.*;
 import leets.leenk.domain.user.domain.entity.User;
+import leets.leenk.global.common.entity.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.time.LocalDateTime;
+
 @Getter
 @Entity
 @SuperBuilder
-@Table(
-        name = "reactions",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"feed_id", "user_id"})
-        }
-)
+@Table(name="feeds")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Reaction {
+public class Feed extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int reactionCount;
+    @Column(length = 100)
+    private String description;
 
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "feed_id", nullable = false, updatable = false)
-    private Feed feed;
+    @Column(nullable = false)
+    private long totalReactionCount;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "user_id", nullable = false, updatable = false)
     private User user;
+
+    private LocalDateTime deletedAt;
+
 }
