+++ src/main/java/leets/leenk/domain/feed/domain/entity/LinkedUser.java
@@ -1,36 +1,33 @@
-package leets.leenk.domain.feed.domain.enitty;
+package leets.leenk.domain.feed.domain.entity;
 
 import jakarta.persistence.*;
 import leets.leenk.domain.user.domain.entity.User;
-import leets.leenk.global.common.entity.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
-import java.time.LocalDateTime;
-
 @Getter
 @Entity
 @SuperBuilder
-@Table(name="feeds")
+@Table(
+        name = "linked_users",
+        uniqueConstraints = {
+                @UniqueConstraint(columnNames = {"feed_id", "user_id"})
+        })
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Feed extends BaseEntity {
+public class LinkedUser {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(length = 100)
-    private String description;
-
-    @Column(nullable = false)
-    private long totalReactionCount;
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "feed_id", nullable = false, updatable = false)
+    private Feed feed;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "user_id", nullable = false, updatable = false)
     private User user;
-
-    private LocalDateTime deletedAt;
 
 }
