+++ src/main/java/leets/leenk/domain/feed/domain/repository/FeedRepository.java
@@ -1,6 +1,6 @@
 package leets.leenk.domain.feed.domain.repository;
 
-import leets.leenk.domain.feed.domain.enitty.Feed;
+import leets.leenk.domain.feed.domain.entity.Feed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FeedRepository extends JpaRepository<Feed, Long> {
