+++ src/main/java/leets/leenk/domain/feed/domain/repository/LinkedUserRepository.java
@@ -1,6 +1,6 @@
 package leets.leenk.domain.feed.domain.repository;
 
-import leets.leenk.domain.feed.domain.enitty.LinkedUser;
+import leets.leenk.domain.feed.domain.entity.LinkedUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LinkedUserRepository extends JpaRepository<LinkedUser, Long> {
