+++ src/main/java/leets/leenk/domain/feed/domain/repository/ReactionRepository.java
@@ -1,6 +1,6 @@
 package leets.leenk.domain.feed.domain.repository;
 
-import leets.leenk.domain.feed.domain.enitty.Reaction;
+import leets.leenk.domain.feed.domain.entity.Reaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReactionRepository extends JpaRepository<Reaction, Long> {
