+++ src/main/java/leets/leenk/domain/feed/domain/service/FeedSaveService.java
@@ -1,6 +1,6 @@
 package leets.leenk.domain.feed.domain.service;
 
-import leets.leenk.domain.feed.domain.enitty.Feed;
+import leets.leenk.domain.feed.domain.entity.Feed;
 import leets.leenk.domain.feed.domain.repository.FeedRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
